+++ samples/react-application-tenant-global-navbar/src/extensions/tenantGlobalNavBar/components/ITenantGlobalFooterBarProps.ts
@@ -1,4 +1,4 @@
-import * as SPTermStore from '../../../components/SPTermStoreService'; 
+import * as SPTermStore from './../services/SPTermStoreService'; 
 
 export interface ITenantGlobalFooterBarProps {
     menuItems: SPTermStore.ISPTermObject[];
